+++ tests/performance/suites/mobilelatency/dependencies_test.go
@@ -1,4 +1,4 @@
-package basicauth
+package mobilelatency
 
 import (
 	"fmt"